+++ fastembed/image/image_embedding.py
@@ -4,14 +4,15 @@ from dataclasses import asdict
 from fastembed.common.types import NumpyArray, Device
 from fastembed.common import ImageInput, OnnxProvider
 from fastembed.image.image_embedding_base import ImageEmbeddingBase
-from fastembed.image.onnx_embedding import OnnxImageEmbedding, NomicVisionEmbedding
+from fastembed.image.onnx_embedding import OnnxImageEmbedding
+from fastembed.image.normalized_embedding import NormalizedEmbedding
 from fastembed.common.model_description import DenseModelDescription
 
 
 class ImageEmbedding(ImageEmbeddingBase):
     EMBEDDINGS_REGISTRY: list[Type[ImageEmbeddingBase]] = [
         OnnxImageEmbedding,
-        NomicVisionEmbedding,
+        NormalizedEmbedding,
     ]
 
     @classmethod
